user: Seyed Vahid Azhari <azharivs@iust.ac.ir>  branch 'Fakhralis'  changed results/plot_results.m  changed src/applications/model/udp-client.cc  changed src/network/model/packet.cc  changed src/network/model/packet.h  changed src/wifi/model/bss-phy-mac-stats.cc  changed src/wifi/model/mac-low.cc  changed src/wifi/model/mpdu-universal-aggregator.cc  changed src/wifi/model/per-sta-q-info.h >>> Bug fix: Strange header fields in A-MPDUs solved!
Update: I realized that Ns3 is unable to extract the right header based on the type of header provided as input to PeekHeader(Header &haedr).
I sure do hope NS3 did include such a capability. Was this not included due to speed optimization? I mean Packet::Print already does this!

Therefore, I introduced another PeekHeader that takes as input the starting index for the search. It is implemented as:

/************************************/
uint32_t
Packet::PeekHeader (Header &header, uint32_t start) const
{
  Buffer::Iterator bufIt = m_buffer.Begin();
  bufIt.Next(start);
  uint32_t deserialized = header.Deserialize (bufIt);
  bufIt.Prev(start);
  NS_LOG_FUNCTION (this << header.GetInstanceTypeId ().GetName () << deserialized);
  return deserialized;
}
/*********************************/

Now I first call PeekHeader for AmpduSubframeHeader and then record the size retured and then do a PeekHeader on WifiMacHeader. I can probably also do a GetSerializedSize() on tha AmpduSubframeHeader and use that as the starting index as well. So my new code looks like:

/*****************************/
.....
    WifiMacHeader hdr;
    AmpduTag ampduTag;
    uint32_t start = 0;
    if (m_curPacket->PeekPacketTag(ampduTag))
      {
        if (ampduTag.GetAmpdu())
          {//extract WifiMacHeader
            AmpduSubframeHeader ampduHdr;
            start = m_curPacket->PeekHeader(ampduHdr);
          }
      }
    m_curPacket->PeekHeader(hdr,start);
.....
/****************************/

However,  it is assumed that correct starting byte index is provided so knowledge of headers before the header of interest should be available. Otherwise unwanted behavior would result.