user: Seyed Vahid Azhari <azharivs@iust.ac.ir>  branch 'source_rate_adapt'  changed sim_run.sh  changed src/wifi/model/per-sta-q-info.cc  changed src/wifi/model/per-sta-q-info.h
>>> I have changed statistics collection to produce time averages instead of sample averages for the following two parameteres:
. average queue size in bytes
. average arrival rate in bytes per second

The averaging interval is controlled from within per-sta-q-info.cc using the Attribute <ObservationInterval>, which is currently set to 2.0 seconds. Other suitable values can be 1.0 and 0.5 seconds. Note that this parameter cannot be set from the main() yet!

This change is made hoping that it eliminates large variations of arrival rate due to instantaneous video frame arrivals, which destabilize the PID controller.